Consolidate utility primitives
